Reduce code duplication between UORBNode and UBXNode
