+++ ros/gisnav/gisnav/extensions/ubx_node.py
@@ -10,7 +10,7 @@ import tf2_geometry_msgs
 import tf2_ros
 import tf_transformations
 from builtin_interfaces.msg import Time
-from geometry_msgs.msg import PointStamped, PoseStamped, TwistWithCovariance, Vector3
+from geometry_msgs.msg import PointStamped, TwistWithCovariance, Vector3
 from nav_msgs.msg import Odometry
 from pyproj import Transformer
 from rcl_interfaces.msg import ParameterDescriptor
@@ -138,249 +138,11 @@ class UBXNode(Node):
     def _publish(self, odometry: Odometry) -> None:
         @narrow_types(self)
         def _publish_inner(odometry: Odometry) -> None:
-            # TODO use inverse publish rate for duration
-            assert odometry.header.frame_id == "gisnav_odom"
-
-            # Do not use timestamp from VO, use latest global match timestamp
-            # instead. This is to avoid interpolating in the gisnav_map frame which
-            # leads to bad results because the observations there are very sparse if
-            # e.g. running on CPU only
-            if self._latest_global_match_stamp is None:
-                self.get_logger().warning(
-                    "No global match timestamps yet, skipping publishing mock GPS msg"
-                )
-                return None
-            odom_to_earth = tf_.lookup_transform(
-                self._tf_buffer,
-                "earth",
-                self._odometry.header.frame_id,
-                (
-                    self._latest_global_match_stamp,
-                    rclpy.duration.Duration(seconds=0.2),
-                ),  # (odometry.header.stamp, rclpy.duration.Duration(seconds=0.2)),
-                logger=self.get_logger(),
-            )
-            if odom_to_earth is None:
-                self.get_logger().warning(
-                    f"Could not determine transform from {odometry.header.frame_id} "
-                    f"to earth"
-                )
-                return None
-
-            odom_time = rclpy.time.Time(
-                seconds=odometry.header.stamp.sec,
-                nanoseconds=odometry.header.stamp.nanosec,
-            )
-            # transform time is same as self._latest_global_match_stamp?
-            transform_time = rclpy.time.Time(
-                seconds=odom_to_earth.header.stamp.sec,
-                nanoseconds=odom_to_earth.header.stamp.nanosec,
-            )
-            try:
-                transform_bridge = self._tf_buffer.lookup_transform_full(
-                    "gisnav_base_link",
-                    transform_time,
-                    "gisnav_base_link",
-                    odom_time,
-                    "gisnav_odom",
-                    rclpy.duration.Duration(seconds=0.2),
-                )
-            except (
-                tf2_ros.LookupException,
-                tf2_ros.ConnectivityException,
-                tf2_ros.ExtrapolationException,
-            ) as e:
-                self.get_logger().warning(
-                    f"Could not get transform to old odom frame: {e}"
-                )
-                return None
-
-            pose = PoseStamped(header=odometry.header, pose=odometry.pose.pose)
-
-            pose = tf_.add_transform_stamped(pose, transform_bridge)
-            pose = tf_.transform_to_pose(pose)
-
-            pose = tf2_geometry_msgs.do_transform_pose(pose.pose, odom_to_earth)
-            # WGS 84 longitude and latitude, and AGL altitude in meters
-            # TODO: this is alt above ellipsoid, not agl
-
-            lon, lat, alt_agl = tf_.ecef_to_wgs84(
-                pose.position.x, pose.position.y, pose.position.z
-            )
-
-            timestamp = tf_.usec_from_header(odometry.header)
-
-            # Heading (yaw := z axis rotation) variance, assume no covariances
-            pose_cov = odometry.pose.covariance.reshape((6, 6))
-            std_dev_c_z = pose_cov[5, 5]
-            h_variance_rad = std_dev_c_z**2
-
-            # WGS 84 ellipsoid and AMSL altitudes
-            altitudes = self._convert_to_wgs84(
-                lat,
-                lon,
-                alt_agl,
-            )
-            if altitudes is not None:
-                alt_ellipsoid, alt_amsl = altitudes
+            mock_gps_dict = tf_.odom_to_typed_dict(self, odometry)
+            if mock_gps_dict is not None:
+                self.nav_pvt(**mock_gps_dict)
             else:
-                return None
-
-            # Make satellites_visible value unrealistic but technically valid to make
-            # GISNav generated mock GPS messages easy to identify. Do not make this
-            # zero because the messages might then get rejected because of too low
-            # satellite count.
-            satellites_visible = np.iinfo(np.uint8).max
-
-            # Pose variance: eph (horizontal error SD) and epv (vertical error SD),
-            # assume no covariances
-            x_var = pose_cov[0, 0]
-            y_var = pose_cov[1, 1]
-            eph = np.sqrt(x_var + y_var)
-            z_var = pose_cov[2, 2]
-            epv = np.sqrt(z_var)
-
-            # 3D velocity
-            twist_with_covariance = odometry.twist
-            twist_with_covariance = self._transform_twist_with_covariance(
-                twist_with_covariance,
-                odometry.header.stamp,
-                "gisnav_base_link",
-                "gisnav_map_ned",
-            )
-            if twist_with_covariance is None:
-                self.get_logger().warning("Could not determine twist covariances")
-                return None
-
-            vel_n_m_s = twist_with_covariance.twist.linear.x
-            vel_e_m_s = twist_with_covariance.twist.linear.y
-            vel_d_m_s = twist_with_covariance.twist.linear.z
-
-            # Heading
-            transform_earth_to_map = tf_.lookup_transform(
-                self._tf_buffer,
-                "gisnav_map",
-                "earth",
-                time_duration=(
-                    odometry.header.stamp,
-                    rclpy.duration.Duration(seconds=0.2),
-                ),
-                logger=self.get_logger(),
-            )
-            pose_map = tf2_geometry_msgs.do_transform_pose(pose, transform_earth_to_map)
-            euler = tf_transformations.euler_from_quaternion(
-                tf_.as_np_quaternion(pose_map.orientation).tolist()
-            )
-            yaw_rad = euler[2]  # ENU frame
-            yaw_rad = -yaw_rad  # NED frame ("heading")
-
-            if yaw_rad < 0:
-                yaw_rad = 2 * np.pi + yaw_rad
-
-            # re-center yaw to [0, 2*pi), it should be at [-pi, pi) before re-centering
-            yaw_rad += np.pi / 2
-            vehicle_yaw_degrees = np.degrees(yaw_rad)
-            vehicle_yaw_degrees = int(vehicle_yaw_degrees % 360)
-            # MAVLink yaw definition 0 := not available
-            vehicle_yaw_degrees = (
-                360 if vehicle_yaw_degrees == 0 else vehicle_yaw_degrees
-            )
-
-            # Speed variance, assume no covariances
-            twist_cov = twist_with_covariance.covariance.reshape((6, 6))
-            # Twist in ENU -> remap to NED here by swapping x and y axes, z axis
-            # inversion should not affect variance
-            vel_n_m_s_var = twist_cov[1, 1]
-            vel_e_m_s_var = twist_cov[0, 0]
-            vel_d_m_s_var = twist_cov[2, 2]
-            s_variance_m_s = vel_n_m_s_var + vel_e_m_s_var + vel_d_m_s_var
-
-            # Course over ground and its variance
-            def _calculate_cog_variance(
-                vel_n_m_s, vel_e_m_s, vel_n_m_s_var, vel_e_m_s_var
-            ) -> float:
-                numerator = (vel_e_m_s_var * vel_n_m_s**2) + (
-                    vel_n_m_s_var * vel_e_m_s**2
-                )
-                denominator = (vel_e_m_s**2 + vel_n_m_s**2) ** 2
-
-                # Calculate the variance of the CoG in radians
-                cog_var = numerator / denominator
-
-                # TODO handle possible exceptions arising from variance exploding at 0
-                #  velocity (as it should)
-                return float(cog_var)
-
-            def _calculate_course_over_ground(
-                east_velocity: float, north_velocity: float
-            ) -> float:
-                """
-                Calculates course over ground from east and north velocities.
-
-                :param east_velocity: The velocity towards the east in meters per
-                    second.
-                :param north_velocity: The velocity towards the north in meters per
-                    second.
-                :return: The course over ground in degrees from the north, in the range
-                    [0, 2 * pi).
-
-                The course over ground is calculated using the arctangent of the east
-                and north velocities. The result is adjusted to ensure it is within
-                the [0, 2 * pi) range.
-                """
-                magnitude = np.sqrt(east_velocity**2 + north_velocity**2)
-
-                if east_velocity >= 0 and north_velocity >= 0:
-                    # top-right quadrant
-                    course_over_ground_radians = np.arcsin(east_velocity / magnitude)
-                elif east_velocity >= 0 and north_velocity < 0:
-                    # bottom-right quadrant
-                    course_over_ground_radians = 0.5 * np.pi + np.arcsin(
-                        -north_velocity / magnitude
-                    )
-                elif east_velocity < 0 and north_velocity < 0:
-                    # bottom-left quadrant
-                    course_over_ground_radians = np.pi + np.arcsin(
-                        -east_velocity / magnitude
-                    )
-                elif east_velocity < 0 and north_velocity >= 0:
-                    # top-left quadrant
-                    course_over_ground_radians = 1.5 * np.pi + np.arcsin(
-                        north_velocity / magnitude
-                    )
-                else:
-                    # todo: this is unreachable?
-                    course_over_ground_radians = 0.0
-
-                return course_over_ground_radians
-
-            # Compute course over ground - pay attention to sine only being
-            # defined for 0<=theta<=90
-            cog_rad = _calculate_course_over_ground(vel_e_m_s, vel_n_m_s)
-
-            # Compute course over ground variance
-            cog_variance_rad = _calculate_cog_variance(
-                vel_n_m_s, vel_e_m_s, vel_n_m_s_var, vel_e_m_s_var
-            )
-
-            self.nav_pvt(
-                int(lat * 1e7),
-                int(lon * 1e7),
-                alt_ellipsoid,
-                alt_amsl,
-                vehicle_yaw_degrees,
-                h_variance_rad,
-                vel_n_m_s,
-                vel_e_m_s,
-                vel_d_m_s,
-                cog_rad,
-                cog_variance_rad,
-                s_variance_m_s,
-                timestamp,
-                eph,
-                epv,
-                satellites_visible,
-            )
+                self.get_logger().warning("Skipping publishing NavPVT")
 
         _publish_inner(odometry)
 
@@ -388,8 +150,8 @@ class UBXNode(Node):
     @ROS.publish(ROS_TOPIC_RELATIVE_NAV_PVT, 10)  # QoSPresetProfiles.SENSOR_DATA.value,
     def nav_pvt(
         self,
-        lat: int,  # todo update to new message definition with degrees, not 1e7 degrees
-        lon: int,  # todo update to new message definition with degrees, not 1e7 degrees
+        lat: int,
+        lon: int,
         altitude_ellipsoid: float,
         altitude_amsl: float,
         yaw_degrees: int,
@@ -398,7 +160,6 @@ class UBXNode(Node):
         vel_e_m_s: float,
         vel_d_m_s: float,
         cog_rad: float,
-        cog_variance_rad: float,
         s_variance_m_s: float,
         timestamp: int,
         eph: float,
